+++ CHANGELOG.md
@@ -6,9 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
+### Added
+* Add support for using [`alloc::vec::Vec`](https://doc.rust-lang.org/alloc/vec/struct.Vec.html)
 ### Changed
 * Due to dependency updates the MSRV has been updated from 1.60 to 1.62. This should only be relevant if you use the `defmt` feature, but we now only test with 1.62 and not older releases, so it's not guaranteed to work otherwise.
 * Update to `heapless:0.8.0`
+### Breaking Changes
+* You must now select either the feature `use_alloc` or `use_heapless` for the crate to compile. Select `use_heapless`
+  to keep the API from the previous release of this crate.
 
 ## [0.1.1] - 2023-01-07
 ### Changed